auto-tagger/pyproject.toml
[project]
name = "auto-tagger"
version = "0.1.0"
description = "A GitHub Action to automatically bump and/or create tags upon push to the default branch, using SemVer formatting."
authors = [
    { name = "notdodo", email = "6991986+notdodo@users.noreply.github.com" },
]
requires-python = ">3.12.0"
readme = "README.md"
dependencies = [
    "pygithub>=2.7.0",
    "pytest>=9.0.2",
    "semver>=3.0.4",
]

[dependency-groups]
dev = ["mypy[faster-cache]>=1.17.1", "ruff>=0.12.7", "pydantic>=2.11.7"]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
output-format = "github"

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = ["ALL"]
ignore = ["D2", "D4", "ANN", "COM812", "ISC001", "ERA001", "PLR0913", "T201"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
case-sensitive = true

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pycodestyle]
max-line-length = 150

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
pretty = true
show_error_codes = true
show_error_context = true
implicit_reexport = true
namespace_packages = true

[tool.pylint."MESSAGES CONTROL"]
persistent = "no"
disable = ["fixme", "line-too-long"]
